Allow weights to be decimals

+++ src/components/WorkoutPage.tsx
@@ -390,7 +390,7 @@ ${exercise.sets.map((set) => `  - ${set.weight} x ${set.reps}`).join('\n')}`
                     InputProps={{
                       inputProps: { min: 0 }
                     }}
-                    onChange={(e) => handleSetChange(exerciseIndex, setIndex, 'weight', parseInt(e.target.value))}
+                    onChange={(e) => handleSetChange(exerciseIndex, setIndex, 'weight', parseFloat(e.target.value))}
                   />
                   <TextField
                     type="number"
